start breaking core into modules
