+++ src/constants.js
@@ -1,0 +1,1 @@
+module.exports.defaultCanvasSize = [1280, 1024];
