+++ src/symbols.js
@@ -1,0 +1,10 @@
+module.exports.computedStyleSymbol = Symbol();
+module.exports.disabledEventsSymbol = Symbol();
+module.exports.elementSymbol = Symbol();
+module.exports.fullscreenElementSymbol = Symbol();
+module.exports.htmlTagsSymbol = Symbol();
+module.exports.mrDisplaysSymbol = Symbol();
+module.exports.optionsSymbol = Symbol();
+module.exports.pointerLockElementSymbol = Symbol();
+module.exports.prioritySymbol = Symbol();
+module.exports.windowSymbol = Symbol();
